Restructure React code
