+++ internal/explorer.point/src/pages/DeployBlog/DeplyBlog.css
@@ -4,11 +4,11 @@
 
 @font-face {
     font-family: Inter;
-    src: url('../assets/fonts/Inter.ttf');
+    src: url('../../assets/fonts/Inter.ttf');
 }
 
 .deploy-container {
-    font-family: Inter;
+    font-family: Inter, Courier, monospace;
     max-width: 1100px;
     margin: auto;
 }
@@ -17,6 +17,6 @@
     width: 100%;
     height: 100%;
     clip-path: polygon(0 0, 100% 0, 100% 100%, 0 100%, 25% 50%);
-    margin-left: 52;
+    margin-left: 52px;
     z-index: -1;
 }
