+++ internal/explorer.point/src/pages/DeployBlog/index.jsx
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { useAppContext } from '../context/AppContext';
+import { useAppContext } from '../../context/AppContext';
 import { useState } from 'react';
 import Alert from 'react-bootstrap/Alert';
 import Button from 'react-bootstrap/Button';
@@ -8,14 +8,15 @@ import Col from 'react-bootstrap/Col';
 import Image from 'react-bootstrap/Image';
 import Carousel from 'react-bootstrap/Carousel';
 import FormControl from 'react-bootstrap/FormControl';
-import image1 from '../assets/blog-images/blogsoftware-1.png';
-import image2 from '../assets/blog-images/blogsoftware-2.png';
-import image3 from '../assets/blog-images/blogsoftware-3.png';
-import image4 from '../assets/blog-images/blogsoftware-4.png';
-import image5 from '../assets/blog-images/blogsoftware-5.png';
-import image6 from '../assets/blog-images/blogsoftware-6.png';
-import './deploy.css';
+import image1 from '../../assets/blog-images/blogsoftware-1.png';
+import image2 from '../../assets/blog-images/blogsoftware-2.png';
+import image3 from '../../assets/blog-images/blogsoftware-3.png';
+import image4 from '../../assets/blog-images/blogsoftware-4.png';
+import image5 from '../../assets/blog-images/blogsoftware-5.png';
+import image6 from '../../assets/blog-images/blogsoftware-6.png';
+import './DeplyBlog.css';
 
+// TODO: do something better than hard-coding
 const VERSION = '0.1';
 const ROOT_DIR_ID =
     '6ce5c3f525128f4173cd4931870b28d05280d66bd206826be1f23242d05c94bb';
@@ -24,7 +25,7 @@ const ROUTES_FILE_ID =
 const CONTRACT_SOURCE_ID =
     '042a2609875d2f5b628896adfc3affec2fd8aaf104f8824918cd94086872dc66';
 
-const blogImagesCarouselData = [
+const BLOG_IMAGES = [
     { caption: 'Create blog posts', img: image1 },
     { caption: 'Share on your wall', img: image2 },
     { caption: 'Blog post preview - 1', img: image3 },
@@ -211,7 +212,7 @@ const DeployBlog = () => {
                     className="position-relative d-flex flex-column align-items-center justify-content-center"
                 >
                     <h4 className="mb-3 text-white">
-                        {blogImagesCarouselData[activeSlideIndex].caption}
+                        {BLOG_IMAGES[activeSlideIndex].caption}
                     </h4>
                     <Carousel
                         interval={2000}
@@ -220,7 +221,7 @@ const DeployBlog = () => {
                         pause={false}
                         onSlide={setActiveSlideIndex}
                     >
-                        {blogImagesCarouselData.map((item) => (
+                        {BLOG_IMAGES.map((item) => (
                             <Carousel.Item key={item.caption}>
                                 <Image
                                     src={item.img}
