+++ internal/explorer.point/src/pages/Final/index.jsx
@@ -1,330 +1,33 @@
-import React, { useRef, useState } from 'react';
+import React from 'react';
 import Container from 'react-bootstrap/Container';
-import Swal from 'sweetalert2';
-import axios from 'axios';
-import Loading from '../components/Loading';
-
-const DEFAULT_ERROR_MESSAGE = 'Something went wrong';
-const MAX_TWEET_SIZE = 280;
-
-const TwitterIcon = (props) => {
-    const { className, style } = props;
-    return (
-        <svg
-            className={className}
-            style={style}
-            stroke="currentColor"
-            fill="currentColor"
-            strokeWidth="0"
-            viewBox="0 0 16 16"
-            height="1em"
-            width="1em"
-            xmlns="http://www.w3.org/2000/svg"
-        >
-            <path d="M15 3.784a5.63 5.63 0 0 1-.65.803 6.058 6.058 0 0 1-.786.68 5.442 5.442 0 0 1 .014.377c0 .574-.061 1.141-.184 1.702a8.467 8.467 0 0 1-.534 1.627 8.444 8.444 0 0 1-1.264 2.04 7.768 7.768 0 0 1-1.72 1.521 7.835 7.835 0 0 1-2.095.95 8.524 8.524 0 0 1-2.379.329 8.178 8.178 0 0 1-2.293-.325A7.921 7.921 0 0 1 1 12.52a5.762 5.762 0 0 0 4.252-1.19 2.842 2.842 0 0 1-2.273-1.19 2.878 2.878 0 0 1-.407-.8c.091.014.181.026.27.035a2.797 2.797 0 0 0 1.022-.089 2.808 2.808 0 0 1-.926-.362 2.942 2.942 0 0 1-.728-.633 2.839 2.839 0 0 1-.65-1.822v-.033c.402.227.837.348 1.306.362a2.943 2.943 0 0 1-.936-1.04 2.955 2.955 0 0 1-.253-.649 2.945 2.945 0 0 1 .007-1.453c.063-.243.161-.474.294-.693.364.451.77.856 1.216 1.213a8.215 8.215 0 0 0 3.008 1.525 7.965 7.965 0 0 0 1.695.263 2.15 2.15 0 0 1-.058-.325 3.265 3.265 0 0 1-.017-.331c0-.397.075-.77.226-1.118a2.892 2.892 0 0 1 1.528-1.528 2.79 2.79 0 0 1 1.117-.225 2.846 2.846 0 0 1 2.099.909 5.7 5.7 0 0 0 1.818-.698 2.815 2.815 0 0 1-1.258 1.586A5.704 5.704 0 0 0 15 3.785z"></path>
-        </svg>
-    );
-};
-
-const CopyIcon = (props) => {
-    const { className, style } = props;
-    return (
-        <svg
-            className={className}
-            style={style}
-            xmlns="http://www.w3.org/2000/svg"
-            width="16"
-            height="16"
-            fill="currentColor"
-            viewBox="0 0 16 16"
-        >
-            <path d="M4 1.5H3a2 2 0 0 0-2 2V14a2 2 0 0 0 2 2h10a2 2 0 0 0 2-2V3.5a2 2 0 0 0-2-2h-1v1h1a1 1 0 0 1 1 1V14a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1V3.5a1 1 0 0 1 1-1h1v-1z" />
-            <path d="M9.5 1a.5.5 0 0 1 .5.5v1a.5.5 0 0 1-.5.5h-3a.5.5 0 0 1-.5-.5v-1a.5.5 0 0 1 .5-.5h3zm-3-1A1.5 1.5 0 0 0 5 1.5v1A1.5 1.5 0 0 0 6.5 4h3A1.5 1.5 0 0 0 11 2.5v-1A1.5 1.5 0 0 0 9.5 0h-3z" />
-        </svg>
-    );
-};
+import Loading from '../../components/Loading';
+import CopyIcon from '../../components/icons/CopyIcon';
+import TwitterIcon from '../../components/icons/TwitterIcon';
+import useFinal, { MAX_TWEET_SIZE } from './useFinal';
 
 const Final = () => {
-    const [identity, setIdentity] = useState('');
-    const [error, setError] = useState('');
-
-    const [loading, setLoading] = useState(false);
-    const [registering, setRegistering] = useState(false);
-
-    const [eligibility, setEligibility] = useState('');
-
-    const [activationCode, setActivationCode] = useState('');
-
-    const [tweetUrl, setTweetUrl] = useState('');
-    const [tweetUrlError, setTweetUrlError] = useState('');
-
-    const [tweetContent, setTweetContent] = useState('');
-    const [tweetContentError, setTweetContentError] = useState('');
-
-    const [openingTwitter, setOpeningTwitter] = useState(false);
-
-    function validateIdentity(identity) {
-        if (identity === '') {
-            setError('Empty identity');
-            return;
-        }
-
-        if (!/^[a-zA-Z0-9_]+?$/.test(identity)) {
-            setError('Special characters are not allowed');
-            return;
-        }
-
-        if (identity.length < 2) {
-            setError('Handle is too short');
-            return;
-        }
-
-        if (identity.length > 16) {
-            setError('Handle is too long');
-            return;
-        }
-
-        return true;
-    }
-
-    function validateTweetUrl(url) {
-        const regex = new RegExp(
-            `^https://twitter.com/${identity}/status/[0-9]+`,
-            'i',
-        );
-
-        if (!regex.test(url)) {
-            setTweetUrlError(
-                'Cannot recognize the URL. It must be a URL of a tweet, and coming from the same identity handle.',
-            );
-            return;
-        }
-
-        setTweetUrlError('');
-
-        return true;
-    }
-
-    const cleanForm = () => {
-        setActivationCode('');
-
-        setTweetContent('');
-        setTweetContentError('');
-
-        setTweetUrl('');
-        setTweetUrlError('');
-
-        setIdentity('');
-        setEligibility('');
-    };
-
-    const cancelRef = useRef();
-    const debounced = useRef(null);
-    const onChangeHandler = (event) => {
-        if (cancelRef.current) {
-            cancelRef.current();
-        }
-        clearTimeout(debounced.current);
-        const identity = event.target.value;
-        cleanForm();
-        if (!validateIdentity(identity)) {
-            return;
-        }
-        setIdentity(identity);
-        debounced.current = setTimeout(async () => {
-            setError('');
-            setLoading(true);
-            axios
-                .get(`/v1/api/identity/isIdentityEligible/${identity}`, {
-                    cancelToken: new axios.CancelToken((c) => {
-                        cancelRef.current = c;
-                    }),
-                    headers: {
-                        'X-Point-Token': `Bearer ${await window.point.point.get_auth_token()}`,
-                    },
-                })
-                .then(({ data }) => {
-                    setLoading(false);
-                    const { eligibility, reason, code } = data.data;
-                    setEligibility(eligibility);
-                    if (reason) {
-                        setError(reason);
-                    }
-
-                    // show an error if the handle is not available
-                    if (!['free', 'tweet'].includes(eligibility) && !reason) {
-                        setError('Handle is not available.');
-                    }
-
-                    if (code) {
-                        setActivationCode(code);
-                        resetTweetContent(code, identity);
-                    }
-                })
-                .catch((thrown) => {
-                    setLoading(false);
-                    if (!axios.isCancel(thrown)) {
-                        setError(DEFAULT_ERROR_MESSAGE);
-                    }
-                });
-        }, 700);
-    };
-
-    const onChangeUrlHandler = (event) => {
-        const url = event.target.value;
-
-        if (!validateTweetUrl(url)) {
-            return;
-        }
-
-        setTweetUrl(url);
-    };
-
-    const validateTweetContent = (content) => {
-        if (content === '') {
-            setTweetContentError('Tweet content cannot be empty');
-            return false;
-        }
-
-        if (!/( |^)#pointnetwork( |$)/g.test(content)) {
-            setTweetContentError('Tweet content must have #pointnetwork');
-            return false;
-        }
-
-        if (!/( |^)#activation( |$)/g.test(content)) {
-            setTweetContentError('Tweet content must have #activation');
-            return false;
-        }
-
-        if (!/( |^)@pointnetwork( |$)/g.test(content)) {
-            setTweetContentError('Tweet content must have @pointnetwork');
-            return false;
-        }
-
-        if (content.length > MAX_TWEET_SIZE) {
-            setTweetContentError('Tweet content is too long');
-            return false;
-        }
-
-        const regex = new RegExp(
-            `https://pointnetwork.io/activation\\?i=0x${activationCode}&handle=${identity}`,
-            'g',
-        );
-
-        if (!regex.test(content)) {
-            setTweetContentError(
-                `Tweet content must have the activation link https://pointnetwork.io/activation?i=0x${activationCode}&handle=${identity}`,
-            );
-            return false;
-        }
-
-        setTweetContentError('');
-        return true;
-    };
-
-    const onChangeTweetContentHandler = (event) => {
-        const newContent = event.target.value;
-        const isValidContent = validateTweetContent(newContent);
-        if (isValidContent) {
-            setTweetContent(newContent);
-        }
-    };
-
-    const resetTweetContent = (code, identity) => {
-        const defaultTweetContent = `Activating my Point Network handle! @pointnetwork https://pointnetwork.io/activation?i=0x${code}&handle=${identity} #pointnetwork #activation`;
-        setTweetContent(defaultTweetContent);
-        setTweetContentError('');
-    };
-
-    const registerHandler = async () => {
-        try {
-            const { isConfirmed } = await Swal.fire({
-                title: 'Are you sure you want to be known as ' + identity + '?',
-                showCancelButton: true,
-                confirmButtonText: 'Sure!',
-            });
-
-            if (!isConfirmed) {
-                return;
-            }
-
-            setRegistering(true);
-            setError('');
-
-            const csrf_token = window.localStorage.getItem('csrf_token');
-
-            const { data } = await axios({
-                url: '/v1/api/identity/register',
-                method: 'POST',
-                contentType: 'application/json; charset=utf-8',
-                dataType: 'json',
-                data: {
-                    identity,
-                    _csrf: csrf_token,
-                    code: activationCode,
-                    url: tweetUrl,
-                },
-                headers: {
-                    'X-Point-Token': `Bearer ${await window.point.point.get_auth_token()}`,
-                },
-            });
-
-            setRegistering(false);
-
-            const { code, success, reason } = data.data;
-
-            if (code) {
-                setActivationCode(code);
-                resetTweetContent(code, identity);
-                return;
-            }
-
-            if (!success) {
-                Swal.fire({
-                    icon: 'error',
-                    text: reason || DEFAULT_ERROR_MESSAGE,
-                });
-                return;
-            }
-            window.location = '/';
-        } catch (error) {
-            setRegistering(false);
-            Swal.fire({
-                title: DEFAULT_ERROR_MESSAGE,
-                text: 'Error: ' + error.message,
-            });
-        }
-    };
-
-    async function openTwitterWithMessage() {
-        setOpeningTwitter(true);
-        const csrf_token = window.localStorage.getItem('csrf_token');
-        const url = `https://twitter.com/intent/tweet?text=${encodeURIComponent(
-            tweetContent,
-        )}`;
-        try {
-            await axios({
-                url: '/v1/api/identity/open',
-                method: 'POST',
-                data: {
-                    url,
-                    _csrf: csrf_token,
-                },
-                headers: {
-                    'X-Point-Token': `Bearer ${await window.point.point.get_auth_token()}`,
-                },
-            });
-            setOpeningTwitter(false);
-        } catch (error) {
-            setOpeningTwitter(false);
-        }
-    }
-
-    const identityAvailable = eligibility === 'free' || eligibility === 'tweet';
-
-    const tweetSize = tweetContent.length;
-
-    const tweetTooBig = tweetSize > MAX_TWEET_SIZE;
+    const {
+        identity,
+        error,
+        loading,
+        registering,
+        eligibility,
+        activationCode,
+        tweetUrl,
+        tweetUrlError,
+        tweetContent,
+        tweetContentError,
+        openingTwitter,
+        identityAvailable,
+        tweetSize,
+        tweetTooBig,
+        onChangeHandler,
+        onChangeUrlHandler,
+        onChangeTweetContentHandler,
+        resetTweetContent,
+        registerHandler,
+        openTwitterWithMessage,
+    } = useFinal();
 
     return (
         <Container className="p-3 text-dark" style={{ color: '#353535' }}>
