+++ internal/explorer.point/src/pages/Identities.jsx
@@ -4,14 +4,14 @@ import { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import InfiniteScroll from 'react-infinite-scroll-component';
 
+const PAGE_SIZE = 100;
+
 export default function Identities() {
     const [identities, setIdentities] = useState([]);
     const [isLoading, setIsLoading] = useState(true);
     const [hasMore, setHasMore] = useState(true);
     const [identitiesLength, setIdentitiesLength] = useState(0);
     const [cursor, setCursor] = useState(0);
-
-    const PAGE_SIZE = 100;
 
     useEffect(() => {
         fetchIdentities();
